+++ R/chains.R
@@ -1,0 +1,32 @@
+#' Extract Chain-Codes from a PDB Object
+
+#' Extract the codes of all chains from a PDB object.
+
+#' \code{chains} is a generic function to extract the ids
+#' of the chains from a PDB object.
+
+
+#' @param x Object of class \code{pdb} or class \code{atoms};
+#' @param \dots arguments passed to methods.
+
+#' @return \code{chains} returns the codes of the individual chains in the molecule.
+
+
+#' @name chains
+#' @export
+chains = function(x, ...)
+	UseMethod("chains")
+
+#' @rdname chains
+#' @export
+chains.atoms = function(x, sorted = TRUE, ...) {
+	ch = unique(x$chainid);
+	if(sorted) ch = sort(ch);
+	return(ch);
+}
+
+#' @rdname chains
+#' @export
+chains.pdb = function(x, ...) {
+	chains(x$atoms);
+}